feat(empty-state): minimal hero, drop starter cards
Users know why they opened a session. The four starter prompts and the
"Working in <cwd>" line added noise without value — replaced with a single
"Ready when you are." line.
